+++ branches/1.0/hybrid.h
@@ -1,6 +1,7 @@
 /*
  * Copyright (C) 2002 Dave Parker
  * Copyright (C) 2003 Carnegie Mellon University
+ * Copyright (C) 2012 Google Inc
  *
  * This file is part of Ymer.
  *
@@ -17,10 +18,9 @@
  * You should have received a copy of the GNU General Public License
  * along with Ymer; if not, write to the Free Software Foundation,
  * Inc., #59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
- *
- * $Id: hybrid.h,v 1.1 2003-11-07 04:24:42 lorens Exp $
  */
-#include <util.h>
+
+#include <cstdio>
 #include <cudd.h>
 #include <odd.h>
 
@@ -52,8 +52,8 @@ struct HDDNode {
     double val;	// for terminals
     HDDKids kids;	// for non-terminals
   } type;
-  int off;
-  int off2;
+  size_t off;
+  size_t off2;
   SparseBit *sb;
   HDDNode *next;
 };
